requirements.txt
# Python 依赖
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0